requirements-test.txt
pytest==5.3.4
pytest-runner==5.2
pytest-flask==0.15.0
pytest-factoryboy==2.0.3
factory_boy==2.12.0